# docs/swagger.yaml
definitions:
  modelsViewApi.BadRequestError:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  modelsViewApi.CrewFormData:
    properties:
      class:
        type: string
      id:
        type: string
      name:
        type: integer
      rating-id:
        type: string
    type: object
  modelsViewApi.CrewInput:
    properties:
      sailNum:
        type: integer
    required:
    - sailNum
    type: object
  modelsViewApi.CrewParticipantAttachInput:
    properties:
      helmsman:
        type: integer
      participantID:
        type: string
    required:
    - participantID
    type: object
  modelsViewApi.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  modelsViewApi.ParticipantFormData:
    properties:
      birthday:
        type: string
      category:
        type: string
      coach:
        type: string
      fio:
        type: string
      gender:
        type: string
      id:
        type: string
    type: object
  modelsViewApi.ParticipantInput:
    properties:
      birthday:
        type: string
      category:
        type: integer
      coach:
        type: string
      fio:
        type: string
      gender:
        type: integer
      id:
        type: string
    required:
    - birthday
    - category
    - coach
    - fio
    type: object
...
info:
  contact: {}
paths:
...
  /api/ratings/{ratingID}/crews:
    get:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список команд успешно получен
          schema:
            items:
              $ref: '#/definitions/modelsViewApi.CrewFormData'
            type: array
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Не найдены рейтинг или команды
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Получить все команды в рейтинге
      tags:
      - Crew
    post:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Данные для создания новой команды
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.CrewInput'
      produces:
      - application/json
      responses:
        "201":
          description: Команда успешно создана
          schema:
            $ref: '#/definitions/modelsViewApi.CrewFormData'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Рейтинг не найден
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Создать новую команду
      tags:
      - Crew
  /api/ratings/{ratingID}/crews/{crewID}:
    delete:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор команды
        format: uuid
        in: path
        name: crewID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Команда успешно удалена
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Рейтинг или команда не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Удалить команду
      tags:
      - Crew
    get:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор команды
        format: uuid
        in: path
        name: crewID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о команде успешно получена
          schema:
            $ref: '#/definitions/modelsViewApi.CrewFormData'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Рейтинг или команда не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Получить информацию о команде
      tags:
      - Crew
    put:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор команды
        format: uuid
        in: path
        name: crewID
        required: true
        type: string
      - description: Данные для обновления команды
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.CrewInput'
      produces:
      - application/json
      responses:
        "200":
          description: Команда успешно обновлена
          schema:
            $ref: '#/definitions/modelsViewApi.CrewFormData'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Рейтинг или команда не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Обновить номер паруса команды
      tags:
      - Crew
  /api/ratings/{ratingID}/crews/{crewID}/members:
    get:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор команды
        format: uuid
        in: path
        name: crewID
        required: true
        type: string
      - description: Фильтр по имени участника
        in: query
        name: name
        type: string
      - description: Фильтр по статусу рулевого (1 – рулевой, 0 – не рулевой)
        in: query
        name: helmsman
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список участников команды
          schema:
            items:
              $ref: '#/definitions/modelsViewApi.ParticipantFormData'
            type: array
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Рейтинг, команда или участники не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Получить список участников команды
      tags:
      - Crew Members
    post:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор команды
        format: uuid
        in: path
        name: crewID
        required: true
        type: string
      - description: Данные для добавления участника
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.CrewParticipantAttachInput'
      produces:
      - application/json
      responses:
        "201":
          description: Участник успешно добавлен
          schema:
            type: string
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Рейтинг или команда не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Добавить участника в команду
      tags:
      - Crew Members
  /api/ratings/{ratingID}/crews/{crewID}/members/{participantID}:
    delete:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор команды
        format: uuid
        in: path
        name: crewID
        required: true
        type: string
      - description: Уникальный идентификатор участника
        format: uuid
        in: path
        name: participantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Участник успешно удален
          schema:
            type: string
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Рейтинг или команда не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Удалить участника из команды
      tags:
      - Crew Members
    get:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор команды
        format: uuid
        in: path
        name: crewID
        required: true
        type: string
      - description: Уникальный идентификатор участника
        format: uuid
        in: path
        name: participantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация об участнике успешно получена
          schema:
            $ref: '#/definitions/modelsViewApi.ParticipantFormData'
        "404":
          description: Команда, рейтинг или участник не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Получить информацию об участнике команды
      tags:
      - Crew Members
    put:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор команды
        format: uuid
        in: path
        name: crewID
        required: true
        type: string
      - description: Уникальный идентификатор участника
        format: uuid
        in: path
        name: participantID
        required: true
        type: string
      - description: Данные для обновления участника
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.ParticipantInput'
      produces:
      - application/json
      responses:
        "200":
          description: Информация об участнике успешно обновлена
          schema:
            $ref: '#/definitions/modelsViewApi.ParticipantFormData'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Команда, рейтинг или участник не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Изменить информацию об участнике команды
      tags:
      - Crew Members
...
swagger: "2.0"